Revert "Added Ideas Requests"
