--- src/models/Ideas.js
@@ -1,22 +1,0 @@
-const mongoose = require("mongoose");
-const Schema = mongoose.Schema;
-const UserSchema = new Schema({
-    name: {
-        type: String,
-        required: true,
-    },
-    title: {
-        type: String,
-        required: true,
-    },
-    message: {
-        type: String,
-        required: true
-    },
-    passphrase: {
-        type: String,
-        required: true
-    }
-});
-
-module.exports = mongoose.model("Idea", UserSchema);
